+++ proton-mail-common/src/actions/messages/unlabel.rs
@@ -1,8 +1,9 @@
 use crate::actions::{filter_responses, ActionError, GenericActionData};
 use crate::datatypes::RollbackItemType;
 use crate::models::Message;
-use proton_action_queue::action::Handler as ActionHandler;
-use proton_action_queue::action::{Action, DefaultVersionConverter, Type};
+use proton_action_queue::action::{
+    Action, DefaultVersionConverter, Handler as ActionHandler, Type,
+};
 use proton_api_core::session::{CoreSession, Session};
 use proton_api_mail::services::proton::ProtonMail;
 use proton_core_common::datatypes::{Id, LocalId, RemoteId};
@@ -10,19 +11,19 @@ use serde::{Deserialize, Serialize};
 use stash::stash::{Interface, Stash, Tether};
 use tracing::error;
 
-/// Action which applies a label to messages.
+/// Action which remove a label from messages.
 #[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct Label(GenericActionData<Message>);
+pub struct Unlabel(GenericActionData<Message>);
 
-impl Label {
-    /// Create a new instance which applies `label_id` to the messages with `message_ids`.
+impl Unlabel {
+    /// Create a new instance which remove `label_id` from the messages with `message_ids`
     pub fn new(label_id: LocalId, message_ids: impl IntoIterator<Item = LocalId>) -> Self {
         Self(GenericActionData::new(label_id, message_ids))
     }
 }
 
-impl Action for Label {
-    const TYPE: Type = Type("label_messages");
+impl Action for Unlabel {
+    const TYPE: Type = Type("unlabel_messages");
     const VERSION: u32 = 1;
     type VersionConverter = DefaultVersionConverter<Self>;
     type Handler = Handler;
@@ -34,7 +35,7 @@ impl Action for Label {
 pub struct Handler;
 
 impl ActionHandler for Handler {
-    type Action = Label;
+    type Action = Unlabel;
 
     async fn apply_local(
         &self,
@@ -42,7 +43,7 @@ impl ActionHandler for Handler {
         tx: &Tether,
     ) -> Result<(), <Self::Action as Action>::Error> {
         action.0.resolve_ids(tx).await?;
-        Message::apply_label(action.0.label_id, action.0.target_ids.clone(), tx).await?;
+        Message::remove_label(action.0.label_id, action.0.target_ids.clone(), tx).await?;
         Ok(())
     }
 
@@ -51,7 +52,7 @@ impl ActionHandler for Handler {
         action: &mut Self::Action,
         tx: &Tether,
     ) -> Result<(), <Self::Action as Action>::Error> {
-        Message::remove_label(action.0.label_id, action.0.target_ids.clone(), tx).await?;
+        Message::apply_label(action.0.label_id, action.0.target_ids.clone(), tx).await?;
         action
             .0
             .mark_rollback(RollbackItemType::Message, tx)
@@ -80,21 +81,21 @@ impl ActionHandler for Handler {
             .expect("Should be set")
             .into();
         let response = api
-            .put_messages_label(message_ids, label_id, None)
+            .put_messages_unlabel(message_ids, label_id)
             .await?
             .responses;
 
         let failed_ids = filter_responses(response);
 
         if !failed_ids.is_empty() {
-            error!("Label messages operation failed for: {failed_ids:?}");
+            error!("Unlabel messages failed for: {failed_ids:?} ");
 
             let tx = stash.transaction().await?;
             let local_ids = RemoteId::counterparts::<Message, _>(failed_ids.clone(), &tx).await?;
 
-            Message::remove_label(action.0.label_id, local_ids, &tx)
+            Message::apply_label(action.0.label_id, local_ids, &tx)
                 .await
-                .inspect_err(|e| error!("Failed to rollback label on messages: {e}"))?;
+                .inspect_err(|e| error!("Failed to rollback unlabel on messages: {e}"))?;
             tx.commit().await?;
         }
         Ok(())