+++ proton-mail-uniffi/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+### Added
+
+  - `apply_label_to_messages` who apply a label to many messages.
+  - `remove_label_from_messages` who remove a label from many messages.
+
 ## [0.11.26] - 2024-09-02
 
 ### Fixed